% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-fn.R
\name{mode}
\alias{mode}
\title{Arithmetic Mode}
\usage{
mode(x)
}
\arguments{
\item{x}{A vector or list.}
}
\value{
The mode of the input.
}
\description{
Compute the arithmetic mode of a set of data. The mode is the most commonly
appearing value.
}
\examples{
mode(c(1, 1, 2))
mode(list(1, 1, 2))
mode(c(1, NA, 1))
}
