% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read}
\alias{read}
\title{Basic function to read in a file}
\usage{
read(
  .file,
  .col_names = FALSE,
  .col_select = everything(),
  .name_repair = "unique"
)
}
\arguments{
\item{.file}{Path to a local file.}

\item{.col_names}{Copied from vroom.}

\item{.col_select}{Copied from vroom.}

\item{.name_repair}{Copied from vroom.}
}
\description{
Basic function to read in a file
}
