% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testactionspkg-package.R
\docType{package}
\name{testactionspkg-package}
\alias{testactionspkg}
\alias{testactionspkg-package}
\title{testactionspkg: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/arisp99/testactionspkg}
  \item Report bugs at \url{https://github.com/arisp99/testactionspkg/issues}
}

}
\author{
\strong{Maintainer}: Aris Paschalidis \email{aris.paschalidis@gmail.com} (\href{https://orcid.org/0000-0003-2247-1885}{ORCID})

}
\keyword{internal}
